Update Grafana volume configuration and add Jenkins service
In the docker-compose.yml file, the Grafana volume configuration was updated to use an environment variable for improved customizability. Additionally, a Jenkins service was added providing capability for continuous integration. A Jenkins container is now created and configured when docker-compose is run.

In the .env file, a new environment variable named GRAFANA_VOLUME was added to be used in the docker-compose.yml for specifying the Grafana storage volume.