Add Prometheus volume to docker-compose and .env files
The change was to add Prometheus volume in docker-compose.yml. This volume links to a local environment variable defined in .env file. This step is necessary for Prometheus to have persistent data storage. It allows Prometheus to store its time series data that is not lost on container restarts. Using a volume claim for storage means your Prometheus data can survive pod restarts, and you can also scale your Prometheus instances up and down as required.