Correct docker-compose relative paths
The paths to files in the docker-compose.yml were corrected to be relative from the root directory. Previously, the paths, particularly to the 'prometheus/grok_exporter' directories, were erroneously referencing absolute paths. This change ensures that the docker-compose command can find the necessary files regardless of the user's current directory when executing the command.