Add monitoring for Expression Cluster pipeline
Added a new grok_exporter service for monitoring Expression Cluster extraction pipeline and set up a corresponding rule in Prometheus.

This change was necessary to track the errors, runs with extracted data, processed papers, and successful runs of the Expression Cluster extraction pipeline. Furthermore, an alert for Expression Cluster Failure was added to Prometheus rule set to ensure proactive error detection. The tracking and alerts allow for improved overview, error detection and therefore increased stability of the pipeline.