Add new grok_exporter_antibody service
Added grok_exporter_antibody service to docker compose file to monitor antibody pipeline. Created a config file and added it as a volume to the service. Additionally added this service as a static config in Prometheus for metrics tracking and added an AntibodyFailure alert to catch errors occurring in the antibody pipeline, providing more granular monitoring of this specific pipeline.