Add Let's Encrypt volume to nginx in docker-compose.yml
In response to recent updates in our server infrastructure, this commit includes modifications to the docker-compose.yml file by adding a new volume to the nginx service. This volume is intended to map the directory where Let's Encrypt stores its SSL certificates on the host machine. This change allows the nginx service running in a Docker container access to these certificates thus enabling HTTPS connections.