Add cron job for automatic SSL certificate renewal
Added a new cron job for automatic SSL certificate renewal with Certbot to mitigate the risk of having an expired SSL certificate. A 'crontab' file has been created to store the cron job and the Dockerfile has been updated accordingly to install cron, set appropriate permissions and also to make sure the cron job is executed when the Docker container starts.