Improve reverse proxy Dockerfile configuration
Updated the Dockerfile in the reverse proxy directory for better maintainability. This update reduces the number of RUN instructions creating fewer layers in the image, and thus reduces the image size. Also, the removal command of cron and tail was shifted to a separate script "entrypoint-wrapper.sh", which is run as the container's main process. The change enables cron to remain running in the background. Cmd is set to keep NGINX up and running.