Correct IP_INSIDE_DOCKER extraction in startup script
Removed unnecessary cut command at the end of IP_INSIDE_DOCKER assignment in the container startup script. It used to take the first part of the ip before '/', which was unnecessary as the awk command was already taking care of retrieving the correct IP address. By removing the cut command, we avoid unnecessary errors and failures during startup if the IP format changes unexpectedly.