Update Dockerfile to add separate nginx conf for prod and dev
The Dockerfile is updated to include different nginx configuration files for production and development environment. Previously, a single configuration was used for both environments which may not be suitable. Container start-up script is updated to copy the relevant nginx configuration file based on the ENV_STATE environment variable. This allows for better control and customisation based on the environment requirements, enhancing the service's flexibility and reliability in different scenarios.